Add nameday command
* move kachna to separate class
